requirements.txt
streamlit
ultralytics
deepface
opencv-python-headless
pillow
tensorflow
tf-keras